+++ Code/ItemEdits/ElectricBoomerang.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using Mono.Cecil.Cil;
+using MonoDetour;
+using MonoDetour.Cil;
+using MonoDetour.DetourTypes;
+using MonoDetour.HookGen;
+using MonoMod.Cil;
+using RoR2;
+
+namespace LordsItemEdits.ItemEdits
+{
+
+    internal static class ElectricBoomerang
+    {
+
+    }
+}