+++ Code/ModSupport/RiskyTweaksMod/MulTScrapLauncherSynergyEdit.cs
@@ -1,4 +1,6 @@
-﻿using LordsItemEdits.ItemEdits;
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
 using Mono.Cecil.Cil;
 using MonoDetour;
 using MonoDetour.Cil;
@@ -7,28 +9,13 @@ using MonoDetour.HookGen;
 using MonoDetour.Logging;
 using MonoMod.Cil;
 using RiskyTweaks.Tweaks.Survivors.Toolbot;
+using LordsItemEdits.ItemEdits;
 using RoR2;
-using System;
-using System.Collections.Generic;
-using System.Text;
 using UnityEngine;
 using HarmonyLib;
 
 namespace LordsItemEdits.ModSupport.RiskyTweaksMod
 {
-    /*[HarmonyPatch]
-    internal static class MulTScrapLauncherSynergyEditHarmony
-    {
-        [HarmonyPatch(typeof(ScrapICBM), nameof(ScrapICBM.FireGrenadeLauncher_ModifyProjectileAimRay))]
-        [HarmonyPrefix]
-        internal static bool PatchItPls(On.EntityStates.Toolbot.FireGrenadeLauncher.orig_ModifyProjectileAimRay orig, ref EntityStates.Toolbot.FireGrenadeLauncher self, ref Ray aimRay, ref Ray __result)
-        {
-            self.damageCoefficient *= PocketICBM.GetICBMDamageMult(self.characterBody);
-            //orig(self, aimRay);
-            return false;
-        }
-    }*/
-
     [MonoDetourTargets(typeof(ScrapICBM), GenerateControlFlowVariants = true)]
     internal static class MulTScrapLauncherSynergyEdit
     {
@@ -40,9 +27,6 @@ namespace LordsItemEdits.ModSupport.RiskyTweaksMod
                 return;
             }
 
-            //var m = DefaultMonoDetourManager.New();
-            //m.LogFilter = MonoDetourLogger.LogChannel.IL;
-            //MonoDetourHooks.RiskyTweaks.Tweaks.Survivors.Toolbot.ScrapICBM.FireGrenadeLauncher_ModifyProjectileAimRay.ControlFlowPrefix(OverrideICBMSynergy, manager: m);
             MonoDetourHooks.RiskyTweaks.Tweaks.Survivors.Toolbot.ScrapICBM.FireGrenadeLauncher_ModifyProjectileAimRay.ILHook(JustDoMyThing);
         }
 
@@ -71,15 +55,6 @@ namespace LordsItemEdits.ModSupport.RiskyTweaksMod
         private static void ChangeICBMSynergy(EntityStates.Toolbot.FireGrenadeLauncher fireGrenadeLauncherState)
         {
             fireGrenadeLauncherState.damageCoefficient *= PocketICBM.GetICBMDamageMult(fireGrenadeLauncherState.characterBody);
-        }
-
-
-        private static ReturnFlow OverrideICBMSynergy(ScrapICBM self, ref On.EntityStates.Toolbot.FireGrenadeLauncher.orig_ModifyProjectileAimRay orig, ref EntityStates.Toolbot.FireGrenadeLauncher fireGrenadeLauncherState, ref Ray aimRay, ref Ray returnValue)
-        {
-            fireGrenadeLauncherState.damageCoefficient *= PocketICBM.GetICBMDamageMult(fireGrenadeLauncherState.characterBody);
-            //returnValue = aimRay;
-            orig(fireGrenadeLauncherState, aimRay);
-            return ReturnFlow.SkipOriginal;
         }
     }
 }