+++ Code/ModSupport/RocketSurvivorGuy/RocketSurvivorGuy.cs
@@ -5,7 +5,7 @@ using System.Text;
 namespace LordsItemEdits.ModSupport.RocketSurvivorGuy
 {
     // dumb naming because the rocket plugin uses 2 of the names i would've used for this
-    internal class RocketSurvivorGuy
+    internal static class RocketSurvivorGuy
     {
         private static bool? _enabled;