+++ Code/ModUtil.cs
@@ -9,6 +9,12 @@ namespace LordsItemEdits
 {
     internal static class ModUtil
     {
+        internal static string GetLangFileLocation(string fileName)
+        {
+            string langFolderPath = System.IO.Path.Combine(System.IO.Path.GetDirectoryName(Plugin.PluginInfo.Location), "Language");
+            return System.IO.Path.Combine(langFolderPath, (fileName + ".language"));
+        }
+
         internal static void LogILError(ILContext il, ILCursor c)
         {
             Log.Error($"COULD NOT IL HOOK {il.Method.Name}");